move collection to its own models folder
