+++ django_project/core/models/__init__.py
@@ -1,1 +1,4 @@
 from core.models.metadata.models import *
+from core.models.collection.models import *
+from core.models.rating.models import *
+from core.models.item.models import *
