+++ django_project/core/models/item/models.py
@@ -31,17 +31,3 @@ class ItemRating(Metadata):
 
 #     def __str__(self):
 #         return self.name
-
-
-class CollectionType(Metadata):
-    name = models.CharField(max_length=128)  # 'farmed', 'wild-caught', etc.
-
-    def __str__(self):
-        return self.name
-
-
-class CollectionMethod(Metadata):
-    name = models.CharField(max_length=128)  # 'gillnets', 'longline', 'recirculating tanks', etc.
-
-    def __str__(self):
-        return self.name
